Project essentials
These files are the documents to support the work done in programming to code for the retail shop. The documents include the pseudocode of the final program, showcasing the plan made and executed while making the program. the document named key elements indicate the modules used in the programming as to show the knowledge gained and applied during the making of program.